unGZipIt function placed in zipper object to avoid code redundancy. Removed some unused code.
