+++ src/main/scala/it/polimi/genomics/importer/CistromeImporter/CistromeTransformer.scala
@@ -1,16 +1,11 @@
 package it.polimi.genomics.importer.CistromeImporter
 
-import it.polimi.genomics.importer.GMQLImporter.{GMQLDataset, GMQLSource, GMQLTransformer}
 import java.io._
-
-import scala.collection.mutable
-import scala.io.Source
 import java.util
-import java.util.zip.GZIPInputStream
 
-import org.apache.commons.compress.archivers.tar.{TarArchiveEntry, TarArchiveInputStream}
+import it.polimi.genomics.importer.GMQLImporter.{GMQLDataset, GMQLSource, GMQLTransformer}
+import org.apache.commons.compress.archivers.tar.TarArchiveInputStream
 import org.apache.commons.compress.compressors.gzip.GzipCompressorInputStream
-import org.apache.log4j.{ConsoleAppender, Level, Logger, PatternLayout}
 import org.slf4j.LoggerFactory
 
 /**
@@ -131,38 +126,5 @@ class CistromeTransformer extends GMQLTransformer {
       entry = tarIn.getNextEntry
     }
   }
-  /**
-    * extracts the gzipFile into the outputPath.
-    *
-    * @param gzipFile   full location of the gzip
-    * @param outputPath full path of destination, filename included.
-    */
-  def unGzipIt(gzipFile: String, outputPath: String): Boolean = {
-    val bufferSize = 1024
-    val buffer = new Array[Byte](bufferSize)
-    var unGzipped = false
-    var timesTried = 0
-    while (timesTried < 4 && !unGzipped) {
-      try {
-        val zis = new GZIPInputStream(new BufferedInputStream(new FileInputStream(gzipFile)))
-        val newFile = new File(outputPath)
-        val fos = new FileOutputStream(newFile)
 
-        var ze: Int = zis.read(buffer)
-        while (ze >= 0) {
-
-          fos.write(buffer, 0, ze)
-          ze = zis.read(buffer)
-        }
-        fos.close()
-        zis.close()
-        unGzipped = true
-      } catch {
-        case e: IOException => timesTried += 1
-      }
-    }
-    if (!unGzipped)
-      logger.error("Couldnt UnGzip the file: " + outputPath)
-    unGzipped
-  }
 }
